feat: decouple assetslist header
