+++ packages/kit/src/views/Wallet/AssetsList/index.tsx
@@ -1,17 +1,11 @@
-/* eslint-disable @typescript-eslint/no-unsafe-call */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import React, { ComponentProps, useCallback, useMemo } from 'react';
 
 import { useFocusEffect, useNavigation } from '@react-navigation/core';
-import { useIntl } from 'react-intl';
 
 import {
-  Box,
   Divider,
   FlatList,
-  IconButton,
   ScrollableFlatListProps,
-  Typography,
   useIsVerticalLayout,
   useUserDevice,
 } from '@onekeyhq/components';
@@ -26,15 +20,14 @@ import { useManageTokens } from '@onekeyhq/kit/src/hooks/useManageTokens';
 import {
   HomeRoutes,
   HomeRoutesParams,
-  ModalRoutes,
   RootRoutes,
   RootRoutesParams,
 } from '@onekeyhq/kit/src/routes/types';
-import { ManageTokenRoutes } from '@onekeyhq/kit/src/views/ManageTokens/types';
 
 import backgroundApiProxy from '../../../background/instance/backgroundApiProxy';
 import { getTokenValues } from '../../../utils/priceUtils';
 
+import AssetsListHeader from './AssetsListHeader';
 import TokenCell from './TokenCell';
 
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
@@ -44,43 +37,6 @@ type NavigationProps = NativeStackNavigationProp<
   RootRoutes.Root
 > &
   NativeStackNavigationProp<HomeRoutesParams, HomeRoutes.ScreenTokenDetail>;
-
-const AssetsListHeaderComponent = ({
-  tokenEnabled,
-}: {
-  tokenEnabled: boolean;
-}) => {
-  const intl = useIntl();
-  const navigation = useNavigation<NavigationProps>();
-  return (
-    <Box
-      flexDirection="row"
-      justifyContent="space-between"
-      alignItems="center"
-      pb={3}
-    >
-      <Typography.Heading>
-        {intl.formatMessage({ id: 'asset__tokens' })}
-      </Typography.Heading>
-      {tokenEnabled ? (
-        <IconButton
-          onPress={() =>
-            navigation.navigate(RootRoutes.Modal, {
-              screen: ModalRoutes.ManageToken,
-              params: { screen: ManageTokenRoutes.Listing },
-            })
-          }
-          size="sm"
-          name="AdjustmentsSolid"
-          type="plain"
-          circle
-        />
-      ) : (
-        <Box w={5} />
-      )}
-    </Box>
-  );
-};
 
 export type IAssetsListProps = Omit<
   ComponentProps<typeof Tabs.FlatList>,
@@ -103,8 +59,6 @@ function AssetsList({
   const { accountTokens, balances, prices } = useManageTokens();
   const { account, network } = useActiveWalletAccount();
   const navigation = useNavigation<NavigationProps>();
-  const { tokenEnabled } = network?.settings ?? { tokenEnabled: false };
-
   const valueSortedTokens = useMemo(
     () =>
       accountTokens.slice().sort((a, b) => {
@@ -181,13 +135,9 @@ function AssetsList({
       ]}
       data={valueSortedTokens}
       renderItem={renderListItem}
-      ListHeaderComponent={
-        ListHeaderComponent ?? (
-          <AssetsListHeaderComponent tokenEnabled={tokenEnabled} />
-        )
-      }
+      ListHeaderComponent={ListHeaderComponent ?? <AssetsListHeader />}
       ItemSeparatorComponent={Divider}
-      ListFooterComponent={ListFooterComponent ?? (() => <Box h={8} />)}
+      ListFooterComponent={ListFooterComponent}
       keyExtractor={(_item: TokenType) => _item.id}
       extraData={isVerticalLayout}
       showsVerticalScrollIndicator={false}
